fix file organization
